Refactor and move Template Variable components
