+++ ui/src/dashboards/components/template_variables/TemplateQueryBuilder.js
